Added confirm to delete,login required

+++ blog/views.py
@@ -3,6 +3,7 @@ from django.http import HttpResponse
 from django.conf import settings
 from blog.models import Post,BlogComment
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 def bloghome(request):
@@ -27,6 +28,7 @@ def postcomment(request):
         messages.success(request,"Comment Posted")
     return redirect(f"/blog/{post.slug}")
 
+@login_required
 def addpost(request):
     if request.method=='POST':
         title=request.POST.get('title')
@@ -37,22 +39,30 @@ def addpost(request):
         messages.success(request,"Posted successfully")
     return render(request,'blog/addPost.html')
 
+@login_required
 def allposts(request):
     post_user=Post.objects.filter(author=request.user)
     return render(request,'blog/allPosts.html',{'post_user':post_user})
 
+@login_required
 def deletepost(request,post_id=None):
-    post_to_delete=Post.objects.get(id=post_id)
-    post_to_delete.delete()
-    messages.success(request,"Post Deleted Successfully")
+    if request.user==Post.author:
+        post_to_delete=Post.objects.get(id=post_id)
+        post_to_delete.delete()
+        messages.success(request,"Post Deleted Successfully")
+    else:
+        return HttpResponse("404 error")
     return redirect("/blog/allposts")
 
-
+@login_required
 def editpost(request,post_id=None):
-    post_to_edit=Post.objects.get(id=post_id)
-    if request.method=='POST':    
-        post_to_edit.title=request.POST.get('title')
-        post_to_edit.content=request.POST.get('content')
-        post_to_edit.save()
-        messages.success(request,"Post Edited")
+    if request.user==Post.author:
+        post_to_edit=Post.objects.get(id=post_id)
+        if request.method=='POST':    
+            post_to_edit.title=request.POST.get('title')
+            post_to_edit.content=request.POST.get('content')
+            post_to_edit.save()
+            messages.success(request,"Post Edited")
+    else:
+        return HttpResponse("404 error")
     return render(request,'blog/editPost.html',{'post_to_edit':post_to_edit})
